+++ CMakeLists.txt
@@ -64,7 +64,6 @@ ament_auto_add_library(${PROJECT_NAME}_hardware SHARED
   src/${PROJECT_NAME}/hardware_model/thruster_direct/esc_direct_model.cpp
   src/${PROJECT_NAME}/hardware_model/thruster_direct/servo_direct_model.cpp
   src/${PROJECT_NAME}/hardware_model/can/vesc_model.cpp
-  src/${PROJECT_NAME}/hardware_model/imu/bno055_model.cpp
 
   src/${PROJECT_NAME}/hardware_model/impl/pigpio.cpp
   src/${PROJECT_NAME}/hardware_model/impl/linux_i2c.cpp
@@ -118,7 +117,6 @@ if(BUILD_TESTING)
   )
   ament_auto_add_gmock(test_hardware_model
     test/cpp/hardware_model/headlights.cpp
-    test/cpp/hardware_model/imu/bno055.cpp
     test/cpp/hardware_model/imu.cpp
     test/cpp/hardware_model/indicator_led.cpp
     test/cpp/hardware_model/can.cpp
